requirements.txt
# OpenAI API
openai==1.58.1

# Video processing
moviepy==1.0.3
pillow==10.4.0

# YouTube API
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0

# HTTP requests
requests==2.32.3
requests-oauthlib==2.0.0

# Instagram (unofficial API)
instagrapi==2.1.2

# Configuration
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1

# CLI
click==8.1.8
rich==13.9.4

# Logging
structlog==24.4.0

# Utilities
python-magic-bin==0.4.14
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-mock==3.14.0
